codehub_backend/requirements.txt
# Core Django
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.9.0

# Authentication & Security
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1

# Image processing
Pillow==10.4.0

# Database
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.8
django-redis==5.4.0

# Task queue
celery==5.4.0

# WebSocket support
channels==4.1.0
channels-redis==4.2.0

# OpenAI integration
openai==1.51.2

# Additional utilities
django-environ==0.11.2
django-filter==24.3
django-health-check==3.18.3

# Development tools
black==24.10.0
flake8==7.1.1
isort==5.13.2